requirements.txt
paho-mqtt==1.4.0
kodi-json==1.0.0
